static cast in surf/bndry ints

+++ src/PDE/Integrate/Boundary.cpp
@@ -390,7 +390,7 @@ bndSurfIntFV(
         std::array< real, 3 >
           gp{{ geoFace(f,4), geoFace(f,5), geoFace(f,6) }};
 
-        auto f_Lid = localFaceId[2*f];
+        auto f_Lid = static_cast< std::size_t >(localFaceId[2*f]);
         auto ref_gp_l = tk::fc_coord[f_Lid];
 
         // Compute the basis functions for the left element
